backend/pyproject.toml
[project]
name = "flypy-typing-backend"
version = "0.1.0"
description = "Formal FastAPI backend for flypy-typing."
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.1",
  "asyncpg>=0.30.0",
  "celery[redis]>=5.4.0",
  "email-validator>=2.2.0",
  "fastapi>=0.115.11",
  "pgvector>=0.3.6",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.8.1",
  "python-jose[cryptography]>=3.4.0",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.39",
  "structlog>=25.2.0",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
ai = [
  "langgraph>=0.3.18",
  "openai>=1.68.2"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
